# tbz-chat-backend/src/main/resources/public/api-docs.yaml
openapi: 3.0.1
info:
  title: TBZ-Chat
  version: 1.0.0
  description: API to create and use chat rooms
  contact:
    name: 'Severin Weigold, Noël Monnerat'
    url: 'https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstleyVEVO'
    email: noreply.tbz.chat@gmail.com
  termsOfService: 'https://pornub.com'
servers:
  - url: 'http://localhost:8080'
...
paths:
...
  /users:
    post:
      summary: Register
      operationId: post-users
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  email:
                    type: string
                  username:
                    type: string
                  chats:
                    type: array
                    items: {}
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
      tags:
        - User
...
  /users/own/password:
    put:
      summary: Change password
      operationId: put-users-own-password
      responses:
        '200':
          description: OK
        '409':
          description: Conflict
          content:
            text/plain:
              schema:
                type: string
                example: Incorrect password
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                oldPassword:
                  type: string
                newPassword:
                  type: string
      tags:
        - User
...
  /users/exists:
    get:
      summary: Check if User exists by email
      tags:
        - User
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      operationId: get-users-exists
      parameters:
        - schema:
            type: string
          in: query
          name: email
          required: true
        - schema:
            type: boolean
          in: query
          name: excludeAuthenticated
...